backend/requirements.txt
groq==0.30.0
langchain_community==0.3.27
langchain_experimental==0.3.4
langchain_openai==0.3.28
numpy==2.3.1
python-dotenv==1.1.1
pytubefix==9.3.0
scikit_learn==1.7.1
youtube_transcript_api==1.2.1
